과제 진행 중 0519

+++ myPage/src/main/java/dalmineks/domain/MyOrderStatus.java
@@ -15,7 +15,7 @@ public class MyOrderStatus {
     private Long id;
 
     private Long itemName;
-    private Long price;
+    private Integer price;
     private String orderStatus;
     private String foodStatus;
     private String deliveryStatus;
